remove golang extension
