--- go/go.mod
@@ -1,5 +1,0 @@
-module github.com/laixintao/flameshow
-
-go 1.18
-
-require github.com/google/pprof v0.0.0-20230912144702-c363fe2c2ed8
